requirements.txt
astroid==1.6.6
awscli==1.18.127
backcall==0.2.0
botocore==1.17.50
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Django==2.0.2
dnspython==1.16.0
docutils==0.15.2
eventlet==0.28.0
Flask==1.1.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.1
git-remote-codecommit==1.15.1
greenlet==0.4.17
idna==2.10
ikp3db==1.4.1
importlib-metadata==1.7.0
Inflector==2.0.12
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
oauthlib==3.1.0
parso==0.1.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
random-username==1.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
spoonacular==3.0
SQLAlchemy==1.3.19
stevedore==3.2.0
traitlets==4.3.3
tweepy==3.9.0
twilio==6.45.1
urllib3==1.25.10
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
